backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.4
pydantic-settings==2.6.1
duckdb==1.3.2
pandas>=2.2.2
scikit-learn>=1.5.2
python-multipart==0.0.9
jinja2==3.1.6
python-jose[cryptography]==3.3.0
structlog==25.4.0
pyyaml==6.0.2
pytest==8.3.2
